refactor: Centralize all zipping logic
makeZipFolder was starting a job in an action and we had to listen in
the parent component to realtime events to know when the job was
finished. Here we wrap all zipping logic in a single function that
returns a promise to separate component logic and zipping logic.
