+++ packages/cozy-mespapiers-lib/src/components/Actions/Items/forwardTo.js
@@ -1,22 +1,13 @@
 import React, { forwardRef } from 'react'
 
-import { getDisplayName } from 'cozy-client/dist/models/contact'
 import ActionsMenuItem from 'cozy-ui/transpiled/react/ActionsMenu/ActionsMenuItem'
 import Icon from 'cozy-ui/transpiled/react/Icon'
 import ListItemIcon from 'cozy-ui/transpiled/react/ListItemIcon'
 import ListItemText from 'cozy-ui/transpiled/react/ListItemText'
 
-import { fetchCurrentUser } from '../../../helpers/fetchCurrentUser'
-import getOrCreateAppFolderWithReference from '../../../helpers/getFolderWithReference'
 import { makeZipFolder } from '../utils'
 
-export const forwardTo = ({
-  t,
-  f,
-  setFileToForward,
-  setIsBackdropOpen,
-  setZipFolder
-}) => {
+export const forwardTo = ({ t, f, setFileToForward, setIsBackdropOpen }) => {
   const label = t('action.forwardTo')
   const icon = 'reply'
 
@@ -31,24 +22,11 @@ export const forwardTo = ({
       } else {
         setIsBackdropOpen(true)
 
-        const currentUser = await fetchCurrentUser(client)
-        const defaultZipFolderName = t('Multiselect.folderZipName', {
-          contactName: getDisplayName(currentUser),
-          date: f(Date.now(), 'YYYY.MM.DD')
-        })
+        const res = await makeZipFolder({ client, docs, t, f })
 
-        const { _id: parentFolderId } = await getOrCreateAppFolderWithReference(
-          client,
-          t
-        )
+        setIsBackdropOpen(false)
 
-        const zipName = await makeZipFolder({
-          client,
-          files: docs,
-          zipFolderName: defaultZipFolderName,
-          dirId: parentFolderId
-        })
-        setZipFolder({ name: zipName, dirId: parentFolderId })
+        setFileToForward(res)
       }
     },
     Component:
